Move schema to server/
