+++ package.json
@@ -7,7 +7,7 @@
     "dev:server": "ts-node-dev --transpile-only --no-notify api",
     "dev:typecheck": "tsc --noEmit --watch",
     "dev": "concurrently --kill-others \"npm run dev:server\" \"npm run dev:typecheck\"",
-    "nexus:reflect": "ts-node --transpile-only api/schema",
+    "nexus:reflect": "ts-node --transpile-only api/server/schema",
     "build:server": "npm run nexus:reflect && ncc build api/index.ts --out .build",
     "build:serverless": "npm run nexus:reflect && ncc build api/app.ts --out .build",
     "migrate:save": "prisma migrate save --experimental",
